+++ src/Engine.cpp
@@ -68,7 +68,7 @@ public:
    
    double speed() const { return mySpeed; }
    IControllerPtr controller() { return shared_from_this(); }
-   double length() const { return model->dimensions().x; }
+   float length() const { return model->dimensions().x; }
 
    // IController interface
    void actOn(Action anAction);
